+++ include/Fireball.h
@@ -8,7 +8,7 @@ class ProjectileManager;
 class Fireball : public BaseProjectile
 {
 public:
-    Fireball(float x, float y, int velX, int velY, SDL_Texture* texture);
+    Fireball(float x, float y, float velX, float velY, SDL_Texture* texture);
 
     void destroy(ProjectileManager& projectileManager) override;
     void render(SDL_Renderer* renderer) override;
